Crate mozjs_collator_glue

Structs

CollationList

Holder for supported collation identifiers.

CollatorOptions

FFI-friendly version of the option bag.

LangColl 🔒

Pair of a language and a collation that is valid for it.

LocaleList

Data holder for enumerating supported locales.

Enums

CollatorCaseFirst

Whether to give case special precedence.

CollatorIgnorePunctuation

Whether to ignore punctuation.

CollatorNumeric

Whether to sort sequences of decimal digits according to the numeric value of the sequence or on a per-character basis.

CollatorSensitivity

Which differences in the strings should lead to differences in collation comparisons.

Constants

ADDITIONAL_LOCALES 🔒

Locales that are needed on the list for 402 compat or suspected Web compat even if not enumerated by ICU4X. (These should be packed in a fancier way to micro optimize binary size, but not annoying supply-chair-reviewers about the requisite crate for now.)

LANG_COLL_COMBINATIONS_SIZE 🔒

If this overflows after an ICU4X update, simply adjust the number.

MAX_LOCALE_LEN 🔒

Currently 10, because the scriptful locales all have two-character language and two-character region, but could grow to 12 with scriptful locales that have three-character language and region. The parameter for ArrayLocale below needs to match.

Statics

LANG_COLL_COMBINATIONS 🔒

For checking if a given language supports a given collation identifier.

Functions

data_locale_to_tiny 🔒

Converts from ICU4X DataLocale to TinyLocaleStr.

get_lang_coll_combinations 🔒

is_supported_collation 🔒

Checks if a given collation is supported for a given language.

list_collations 🔒

List the supported collation identifiers excluding standard and search.

list_locales 🔒

List the supported locales by merging locales enumerated by ICU4X and additional locales that ICU4X doesn’t enumerate but SpiderMonkey needs to see enumerated.

mozilla_collator_glue_collation_list_free⚠

Free the collation identifier list received from mozilla_collator_glue_collation_list_new.

mozilla_collator_glue_collation_list_item⚠

Read an items from the collation identifier list.

mozilla_collator_glue_collation_list_len⚠

Get the length of the collation identifier list.

mozilla_collator_glue_collation_list_new⚠

List the supported collation identifiers for FFI. The return value must be freed with mozilla_collator_glue_collation_list_free.

mozilla_collator_glue_collator_compare_latin1⚠

Compares Latin1 strings.

mozilla_collator_glue_collator_compare_latin1_utf16⚠

Compares Latin1 string to a UTF-16 string.

mozilla_collator_glue_collator_compare_utf16⚠

Compares UTF-16 strings.

mozilla_collator_glue_collator_free⚠

Deleter for values previously obtained from mozilla_collator_glue_collator_try_new.

mozilla_collator_glue_collator_resolved_options⚠

Returns the resolved options of the collator.

mozilla_collator_glue_collator_try_new⚠

Constructor for the collator.

mozilla_collator_glue_is_supported_collation⚠

Checks if a given collation is supported for a given language.

mozilla_collator_glue_locale_list_free⚠

Free the locale list returned by mozilla_collator_glue_locale_list_new.

mozilla_collator_glue_locale_list_item⚠

Read an item from the locale list.

mozilla_collator_glue_locale_list_len⚠

Get the length of the locale list.

mozilla_collator_glue_locale_list_new⚠

List supported locales for FFI. The return value must be freed using mozilla_collator_glue_locale_list_free.

Type Aliases

ArrayLocale 🔒

Write sink that can fit locales we know about. Parameter needs to match MAX_LOCALE_LEN.

TinyCollationStr 🔒

Tiny string holding a collation type. 8 is the maximum length for a BCP-47 subtag. The full length is used by phonetic.

TinyLocaleStr 🔒

Tiny string for any locale we know about